refactor(routes): extract upload ingest routes
